batch_data_ingestion/requirements.txt
pyspark==3.5.0
mysql-connector-python==8.2.0
boto3==1.34.1
python-dotenv==1.0.0
setuptools>=65.5.1
pandas>=1.5.3
